--- device/src/bsp/boards/mod.rs
@@ -1,6 +1,0 @@
-pub mod stm32h7;
-pub mod stm32l0;
-pub mod stm32l1;
-pub mod stm32l4;
-pub mod stm32u5;
-pub mod stm32wl;
