--- device/src/bsp/boards/nrf52/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "bsp+adafruit_feather_nrf52840")]
-pub mod adafruit_feather_nrf52840;
