--- device/src/bsp/boards/stm32h7/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "bsp+nucleo_h743zi")]
-pub mod nucleo_h743zi;
