--- device/src/bsp/boards/stm32l0/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "bsp+b_l072z_lrwan1")]
-pub mod lora_discovery;
