--- device/src/bsp/boards/stm32l1/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "bsp+rak811")]
-pub mod rak811;
