--- device/src/bsp/boards/stm32l4/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(any(feature = "bsp+b_l4s5i_iot01a", feature = "bsp+b_l475e_iot01a"))]
-pub mod iot01a;
