--- device/src/bsp/boards/stm32u5/b_u585i_iot02a.rs
@@ -1,45 +1,0 @@
-use crate::{
-    bsp::Board,
-    drivers::{
-        button::Button,
-        led::{ActiveHigh, ActiveLow, Led},
-    },
-};
-use embassy_stm32::{
-    exti::ExtiInput,
-    gpio::{Input, Level, Output, Pull, Speed},
-    peripherals::{PC13, PE13, PH6, PH7},
-};
-
-pub type PinLedBlue = Output<'static, PE13>;
-pub type LedBlue = Led<PinLedBlue, ActiveHigh>;
-
-pub type PinLedGreen = Output<'static, PH7>;
-pub type LedGreen = Led<PinLedGreen, ActiveLow>;
-
-pub type PinLedRed = Output<'static, PH6>;
-pub type LedRed = Led<PinLedRed, ActiveLow>;
-
-pub type PinUserButton = Input<'static, PC13>;
-pub type UserButton = Button<ExtiInput<'static, PC13>>;
-
-pub struct Iot02a {
-    pub led_blue: LedBlue,
-    pub led_green: LedGreen,
-    pub led_red: LedRed,
-    pub user_button: UserButton,
-}
-
-impl Board for Iot02a {
-    type Peripherals = embassy_stm32::Peripherals;
-    type BoardConfig = ();
-
-    fn new(p: Self::Peripherals) -> Self {
-        Self {
-            led_blue: Led::new(Output::new(p.PE13, Level::High, Speed::Low)),
-            led_green: Led::new(Output::new(p.PH7, Level::High, Speed::Low)),
-            led_red: Led::new(Output::new(p.PH6, Level::High, Speed::Low)),
-            user_button: Button::new(ExtiInput::new(Input::new(p.PC13, Pull::Down), p.EXTI13)),
-        }
-    }
-}
