--- device/src/bsp/boards/stm32u5/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "bsp+b_u585i_iot02a")]
-pub mod b_u585i_iot02a;
