--- device/src/bsp/boards/stm32wl/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "bsp+nucleo_wl55")]
-pub mod nucleo_wl55;
