+++ examples/blinky/Cargo.toml
@@ -7,10 +7,10 @@ description = "Blinky App"
 [workspace]
 
 [dependencies]
-embassy-executor = { version = "0.1.0", default-features = false, features = ["nightly", "integrated-timers"] }
-embassy-stm32 = { version = "0.1.0", optional = true, features = ["exti", "unstable-traits", "nightly", "time-driver-any"] }
-embassy-nrf = {version = "0.1.0", optional = true, features = ["unstable-traits", "nightly", "gpiote", "time-driver-rtc1"] }
-cortex-m = { version = "0.7.6", features = ["critical-section-single-core"] }
+embassy-executor = { version = "0.1.0", default-features = false, features = ["nightly"] }
+embassy-stm32 = { version = "0.1.0", optional = true, features = ["exti", "unstable-traits", "nightly"] }
+embassy-nrf = {version = "0.1.0", optional = true, features = ["unstable-traits", "nightly", "gpiote"] }
+cortex-m = { version = "0.7", features = ["critical-section-single-core"] }
 cortex-m-rt = "0.7"
 
 embedded-hal = "1.0.0-alpha.8"
@@ -20,11 +20,11 @@ panic-probe = { version = "0.3", features = ["print-defmt"], optional = true }
 defmt-rtt = { version = "0.3", optional = true }
 defmt = "0.3"
 
-drogue-metaboard = { path = "../metaboard", default-features = false }
+drogue-device = { path = "../../device", default-features = false }
 
 [features]
-"board+b-l475e-iot01a" = ["embassy-stm32/stm32l475vg", "panic-probe", "defmt-rtt", "drogue-metaboard/board+b-l475e-iot01a"]
-"board+nrf52-dk" = ["embassy-nrf/nrf52832", "panic-probe", "defmt-rtt", "drogue-metaboard/board+nrf52-dk"]
+"board+b-l475e-iot01a" = ["embassy-stm32/stm32l475vg", "panic-probe", "defmt-rtt", "drogue-device/board+b-l475e-iot01a"]
+"board+nrf52-dk" = ["embassy-nrf/nrf52832", "panic-probe", "defmt-rtt", "drogue-device/board+nrf52-dk"]
 
 [patch.crates-io]
 embassy-executor = { git = "https://github.com/embassy-rs/embassy.git", tag = "embassy-time-v0.1.0" }
