--- examples/device/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "drogue-device"
-version = "0.1.0"
-
-[workspace]
-
-[dependencies]
-drogue-device-macros = { path = "../macros" }
-panic-probe = { version = "0.3", features = ["print-defmt"], optional = true }
-panic-reset = { version = "0.1.1", optional = true }
-defmt-rtt = { version = "0.3.2", optional = true }
-embassy-executor = { git = "https://github.com/embassy-rs/embassy.git", tag = "embassy-time-v0.1.0", default-features = false, features = ["nightly"] }
-embassy-stm32 = { git = "https://github.com/embassy-rs/embassy", tag = "embassy-time-v0.1.0", optional = true }
-embassy-nrf = { git = "https://github.com/embassy-rs/embassy.git", tag = "embassy-time-v0.1.0", optional = true }
-
-[features]
-"board+b_l475e_iot01a" = ["embassy-stm32/stm32l475vg", "dep:panic-probe", "dep:defmt-rtt"]
